improve swiper initialization. separate loaders for different swipers
